overhaul gui by switching to dearpygui library
